refactor: new category dialog to separate component
