+++ desk/src/components/desk/knowledge_base/TopNavbar.vue
@@ -15,7 +15,7 @@
 					<Dropdown :options="action.dropdown" placement="right">
 						<template v-slot="{ toggle }">
 							<div class="py-[4px]" @click="toggle">
-								<FeatherIcon name="chevron-down" class="w-4 h-4 stroke-blue-400" role="button" />
+								<FeatherIcon name="chevron-down" class="w-4 h-4 stroke-white" role="button" />
 							</div>
 						</template>
 					</Dropdown>
